Moved labels to misc
