+++ internal/config/duration_tracker.go
@@ -2,14 +2,15 @@ package config
 
 import (
 	"fmt"
+	"rampart/internal/misc"
 )
 
 type DurationTracker struct {
-	Name      string     `yaml:"name"`
-	Help      string     `yaml:"help"`
-	Buckets   []float64  `yaml:"buckets"`
-	Labels    []string   `yaml:"labels"`
-	Processes *Processes `yaml:"processes"`
+	Name      string          `yaml:"name"`
+	Help      string          `yaml:"help"`
+	Buckets   []float64       `yaml:"buckets"`
+	Labels    []string        `yaml:"labels"`
+	Processes *misc.Processes `yaml:"processes"`
 }
 
 func (tracker *DurationTracker) String() string {
