+++ internal/config/run_tracker.go
@@ -2,13 +2,14 @@ package config
 
 import (
 	"fmt"
+	"rampart/internal/misc"
 )
 
 type RunTracker struct {
-	Name     string    `yaml:"name"`
-	Help     string    `yaml:"help"`
-	Labels   []string  `yaml:"labels"`
-	Statuses *Statuses `yaml:"statuses"`
+	Name     string         `yaml:"name"`
+	Help     string         `yaml:"help"`
+	Labels   []string       `yaml:"labels"`
+	Statuses *misc.Statuses `yaml:"statuses"`
 }
 
 func (tracker *RunTracker) String() string {
