+++ internal/config/storing_tracker.go
@@ -2,13 +2,14 @@ package config
 
 import (
 	"fmt"
+	"rampart/internal/misc"
 )
 
 type StoringTracker struct {
-	Name         string        `yaml:"name"`
-	Help         string        `yaml:"help"`
-	Labels       []string      `yaml:"labels"`
-	Consequences *Consequences `yaml:"consequences"`
+	Name         string             `yaml:"name"`
+	Help         string             `yaml:"help"`
+	Labels       []string           `yaml:"labels"`
+	Consequences *misc.Consequences `yaml:"consequences"`
 }
 
 func (tracker *StoringTracker) String() string {
