+++ internal/misc/verdicts.go
@@ -1,4 +1,4 @@
-package config
+package misc
 
 import (
 	"fmt"
